+++ CHANGELOG.md
@@ -16,14 +16,19 @@ CHANGELOG
 
 ### TODO 1.2.0
 
-- [ ] children_prefix功能,单元测试
+- [ ] children_prefix功能
+- [ ] 单元测试: 父节点连接注册信息的子网范围不包含本节点或本节点的子网，出错
+- [ ] 单元测试: 父子节点在不同的子网范围
+- [ ] 单元测试: 转发目标和来源ep一样，回环，出错
+- [ ] 单元测试: endpoint::merge_subnets
+- [ ] 单元测试: endpoint_subnet_range::contain
+- [ ] 单元测试: endpoint_subnet_range::search_subnet_for_id(====#^++####====)
 - [x] access_token验证的单元测试（成功+失败）
 - [ ] 重构父节点范围管理的结构
 - [x] 协议版本兼容性单元测试
 - [x] 内存通道版本兼容性单元测试
 - [x] 字符串路径共享内存测试
 - [x] 握手connection的backlog支持，定时器优化
-- [ ] unit test -- no router : EN_ATBUS_ERR_ATNODE_TTL
 - [x] unit test -- max msg size : EN_ATBUS_ERR_INVALID_SIZE
 - [x] msg size -> configure: 256KB